Refactor db.ts to break circular dependencies by introducing DbDeps type and setDbDependencies function; update imports in related files for consistency. Adjust layer definitions in layers.ts for qWater to use vector tiles and add new line layers. Clean up file-list component structure by consolidating types into file-list-types.ts.
