+++ website/src/lib/components/toolbar/tools/scissors/Scissors.svelte
@@ -1,9 +1,5 @@
 <script lang="ts" context="module">
-	export enum SplitType {
-		FILES = 'files',
-		TRACKS = 'tracks',
-		SEGMENTS = 'segments'
-	}
+	export { SplitType } from './SplitType';
 </script>
 
 <script lang="ts">
