+++ website/src/lib/components/toolbar/tools/scissors/SplitType.ts
@@ -1,0 +1,6 @@
+/** Split type for scissors tool. In separate file to avoid circular dependency with db. */
+export enum SplitType {
+    FILES = 'files',
+    TRACKS = 'tracks',
+    SEGMENTS = 'segments'
+}
